+++ src/chromatix/utils/windows.py
@@ -3,7 +3,7 @@ import flax.linen as nn
 from chex import Array
 from scipy.ndimage import distance_transform_edt  # type: ignore
 from typing import Tuple
-from ..utils import _broadcast_2d_to_spatial
+from . import _broadcast_2d_to_spatial
 
 
 def sigmoid_taper(shape: Tuple[int, int], width: float, ndim: int = 5) -> Array:
